remove: mainItem, MainViewModel 삭제, feature: mainViewModel -> homeViewModel 함수 이동
